Move entity to domain
