+++ src/domain/auth/entities/jwt-access-token.payload.entity.ts
