+++ src/domain/jwt/jwt.repository.interface.ts
@@ -1,7 +1,7 @@
 import { JwtModule } from '@/datasources/jwt/jwt.module';
 import { JwtRepository } from '@/domain/jwt/jwt.repository';
 import { Module } from '@nestjs/common';
-import { JwtAccessTokenPayload } from '@/routes/auth/entities/jwt-access-token.payload.entity';
+import { JwtAccessTokenPayload } from '@/domain/auth/entities/jwt-access-token.payload.entity';
 import { JwtPayloadWithClaims } from '@/datasources/jwt/jwt-claims.entity';
 
 export const IJwtRepository = Symbol('IJwtRepository');
